+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
+
+## [1.0.33] 2021-11-23
+
+### Added
+- add `PLC_LeakSensor`
+- add `PLC_Fan` (beta)
+### Changed
+- `PLC_HumidifierDehumidifier` add separated bit on/off support by adding  `set_Deactivate`
+- `PLC_HumidifierDehumidifier` add byte support for rotation speed by adding `set_RotationSpeedByte` and  `get_RotationSpeedByte`
+
 ## [1.0.32] 2021-11-11
 
 With the latest update of the Homebridge UI docker container the the plugin does not start!